MAINT: Move EquationNames enum
