+++ src/FTHM_Solver/schemes.py
@@ -58,29 +58,6 @@ essentially bearers of options for the solver.
 """
 
 
-class EquationNames(Enum):
-    """Enum for the names of the equations in the model."""
-
-    MASS_BALANCE = "mass_balance_equation"
-    MASS_BALANCE_MATRIX = "mass_balance_equation"
-    MASS_BALANCE_FRACTURES = "mass_balance_equation"
-    MASS_BALANCE_INTERSECTIONS = "mass_balance_equation"
-    ENERGY_BALANCE = "energy_balance_equation"
-    ENERGY_BALANCE_MATRIX = "energy_balance_equation"
-    ENERGY_BALANCE_FRACTURES = "energy_balance_equation"
-    ENERGY_BALANCE_INTERSECTIONS = "energy_balance_equation"
-    INTERFACE_DARCY_FLUX = "interface_darcy_flux_equation"
-
-    INTERFACE_ENTHALPY_FLUX = "interface_enthalpy_flux_equation"
-    INTERFACE_FOURIER_FLUX = "interface_fourier_flux_equation"
-
-    MECHANICS = "momentum_balance_equation"
-    INTERFACE_FORCE_BALANCE = "interface_force_balance_equation"
-    CONTACT = "contact_mechanics_equation"
-    CONTACT_NORMAL = "normal_fracture_deformation_equation"
-    CONTACT_TANGENTIAL = "tangential_fracture_deformation_equation"
-
-
 def transform_contact_block(J, row_group: int, col_group: int, nd: int):
     """Assemble the right linear transformation."""
     # Sorted according to groups. If not done, the matrix can be in porepy order,
